Added new properties to CrossLink objects
to enable CrossLink weight calculations based on the TYPE of node, which
is not identifable by the previous collection of properties, which only
included ids, which were not enough to make this task possible.

There might be errors! Have not tested it yet!
But the new properties should be passed to Neo4J when the scenario is
built, and should be extracted when the CTAPTransportLinks are created.

Next step is to use these properties in the CTAPTransportLink creation.I
believe this needs to happen in DefaultCTAPTransportLinkFactory