Added more types of nodes to core.graph.Utils.getNodeTypeAsString
because there are apparently many. If it isn't one of the ones I know
about, it now puts "other" instead of throwing an exception.

Also the deleteScenarioPipeline and ScenarioPipeline have everything
commented out but the stuff related to CrossLinks so I can test the new
CrossLink code. Currently running into problems because my VM is too
small (only 30GB of storage space). Will need to get this stuff up and
running in Windows to avoid this problem.