now CrossLink function for ground to air and air to ground implemented
in particular, the air-rail and rail-air links incorporate the effects
of schedule optimization from ENAC. BUT BEWARE: currently the values for
Madrid are used for ALL AIRPORTS, THERE IS NO DIFFERENCE BETWEEN
AIRPORTS.

Also, everything is hard coded into TransitCTAPTransportLinkFactory1,
because we don't have time to figure out how to pass all this optimially
through the rest of the classes....so it might be slow! And you need to
go to that class to change it!

The default are the values BEFORE optimization! There is a switch called
"beforeSchOpt": if this is true, values for before optimization are use.
if false, values for after optimization are used!